Moved to github and fixed build error

+++ tcp/tcb.go
@@ -107,7 +107,7 @@ func (c *TCB) Recv(num int) ([]byte, error) { // blocking recv call TODO add tim
 	for {
 		/*logs*/ logs.Trace.Println(c.hash(), "Attempting to read off of pushBuffer")
 		/*logs*/ logs.Trace.Println(c.hash(), "Amt of data on pushBuffer:", len(c.pushBuffer))
-		amt := min(num, uint64(len(c.pushBuffer)))
+		amt := min(uint64(num), uint64(len(c.pushBuffer)))
 		if amt != 0 {
 			data := c.pushBuffer[:amt]
 			c.pushBuffer = c.pushBuffer[amt:]
